Refactor the Random plugin.
